fixed adapter/migrator imports, still a hack
